requirements/requirements.txt
azure-core==1.34.0
azure-identity==1.23.0
azure-storage-blob==12.25.1
azure-storage-file-datalake==12.20.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
idna==3.10
isodate==0.7.2
msal==1.32.3
msal-extensions==1.3.1
numpy==2.3.2
pandas==2.3.1
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.4
six==1.17.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
